+++ inst/shiny/server.R
@@ -160,14 +160,17 @@ function(input, output, session) {
     }
     x_range <- if (input$userxzoom) c(input$lowerxin, input$upperxin) else NULL
     ref_value <- if (is.na(input$refvalue)) 1 else input$refvalue
-
+    inputxaxistitle <- if(input$parsexaxistitle) parse(text=input$xaxistitle) else input$xaxistitle
+    inputyaxistitle <- if(input$parseyaxistitle) parse(text=input$yaxistitle) else input$yaxistitle
+    
     metaExpr({
       summarydata <- ..(plotdataprepare())
+        
       plot <- forest_plot(
         data = summarydata,
         facet_formula = ..(input$facetformula),
-        xlabel = ..(input$xaxistitle),
-        ylabel = ..(input$yaxistitle),
+        xlabel = ..(inputxaxistitle),
+        ylabel = ..(inputyaxistitle),
         x_facet_text_size = ..(input$facettextx),
         y_facet_text_size = ..(input$facettexty),
         x_facet_text_angle = ..(input$facettextxangle),