+++ NEWS.md
@@ -3,6 +3,7 @@
 * minor ui improvement and control for the facets x and y text hjust and vjust
 * added vignette on pediatric multivariate covariate simulations
 * added facet_labeller argument (function only not in the shiny app)
+* added possibility to supply xlabel and ylabel expressions
 
 # coveffectsplot 0.0.5